+++ src/test/java/org/folio/rest/api/ServicePointTest.java
@@ -1,5 +1,7 @@
 package org.folio.rest.api;
 
+import static org.folio.rest.impl.ServicePointAPI.SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_BEING_PICKUP_LOC;
+import static org.folio.rest.impl.ServicePointAPI.SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_HOLD_EXPIRY;
 import static org.folio.rest.support.http.InterfaceUrls.servicePointsUrl;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.notNullValue;
@@ -14,6 +16,9 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import io.vertx.core.json.Json;
+import io.vertx.core.json.JsonArray;
+import org.folio.rest.jaxrs.model.HoldShelfExpiryPeriod;
 import org.folio.rest.support.AdditionalHttpStatusCodes;
 import org.folio.rest.support.Response;
 import org.folio.rest.support.ResponseHandler;
@@ -35,7 +40,7 @@ import io.vertx.core.logging.LoggerFactory;
  *
  * @author kurt
  */
-public class ServicePointTest {
+public class ServicePointTest extends TestBase{
   private static Logger logger = LoggerFactory.getLogger(ServicePointTest.class);
   private static final String SUPPORTED_CONTENT_TYPE_JSON_DEF = "application/json";
 
@@ -45,8 +50,8 @@ public class ServicePointTest {
           ExecutionException,
           TimeoutException,
           MalformedURLException {
-    StorageTestSuite.deleteAll(servicePointsUrl(""));
-  }
+             StorageTestSuite.deleteAll(servicePointsUrl(""));
+      }
 
   // --- BEGIN TESTS --- //
 
@@ -58,7 +63,8 @@ public class ServicePointTest {
           MalformedURLException {
 
     Response response = createServicePoint(null, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20,
+        true, createHoldShelfExpiryPeriod());
     assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
     assertThat(response.getJson().getString("id"), notNullValue());
     assertThat(response.getJson().getString("code"), is("cd1"));
@@ -72,7 +78,8 @@ public class ServicePointTest {
           TimeoutException,
           MalformedURLException {
     Response response = createServicePoint(null, null, "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null,
+        20, true, createHoldShelfExpiryPeriod());
     assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
   }
 
@@ -83,7 +90,8 @@ public class ServicePointTest {
           TimeoutException,
           MalformedURLException {
     Response response = createServicePoint(null, "Circ Desk 1", null,
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null,
+        20, true, createHoldShelfExpiryPeriod());
     assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
   }
 
@@ -94,7 +102,7 @@ public class ServicePointTest {
           TimeoutException,
           MalformedURLException {
     Response response = createServicePoint(null, "Circ Desk 1", "cd1",
-        null, null, 20, true);
+        null, null, 20, true, createHoldShelfExpiryPeriod());
     assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
   }
 
@@ -105,9 +113,9 @@ public class ServicePointTest {
           TimeoutException,
           MalformedURLException {
     createServicePoint(null, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     Response response = createServicePoint(null, "Circ Desk 1", "cd2",
-        "Circulation Desk -- Bathroom", null, 20, true);
+        "Circulation Desk -- Bathroom", null, 20, true, createHoldShelfExpiryPeriod());
     assertThat(response.getStatusCode(), is(AdditionalHttpStatusCodes.UNPROCESSABLE_ENTITY));
   }
 
@@ -119,12 +127,16 @@ public class ServicePointTest {
           MalformedURLException {
     UUID id = UUID.randomUUID();
     createServicePoint(id, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     Response response = getById(id);
     assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_OK));
     assertThat(response.getJson().getString("id"), is(id.toString()));
     assertThat(response.getJson().getString("code"), is("cd1"));
     assertThat(response.getJson().getString("name"), is("Circ Desk 1"));
+
+    JsonObject holdShelfExpiryPeriod = response.getJson().getJsonObject("holdShelfExpiryPeriod");
+    assertThat(holdShelfExpiryPeriod.getInteger("duration"), is (2));
+    assertThat(holdShelfExpiryPeriod.getString("durationUnit"), is (HoldShelfExpiryPeriod.DurationUnit.DAYS.toString()));
   }
 
   @Test
@@ -135,9 +147,9 @@ public class ServicePointTest {
     MalformedURLException {
 
     createServicePoint(null, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     createServicePoint(null, "Circ Desk 2", "cd2",
-        "Circulation Desk -- Basement", null, 20, true);
+        "Circulation Desk -- Basement", null, 20, true, createHoldShelfExpiryPeriod());
     CompletableFuture<Response> getCompleted = new CompletableFuture<>();
     send(servicePointsUrl("/"), HttpMethod.GET,
       null, SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(getCompleted));
@@ -155,9 +167,9 @@ public class ServicePointTest {
     MalformedURLException {
 
     createServicePoint(null, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     createServicePoint(null, "Circ Desk 2", "cd2",
-        "Circulation Desk -- Basement", null, 20, true);
+        "Circulation Desk -- Basement", null, 20, true, createHoldShelfExpiryPeriod());
     CompletableFuture<Response> getCompleted = new CompletableFuture<>();
     send(servicePointsUrl("/?query=code==cd1"), HttpMethod.GET,
       null, SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.json(getCompleted));
@@ -175,7 +187,7 @@ public class ServicePointTest {
           MalformedURLException {
     UUID id = UUID.randomUUID();
     createServicePoint(id, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     JsonObject request = new JsonObject()
             .put("id", id.toString())
             .put("name", "Circ Desk 2")
@@ -202,7 +214,7 @@ public class ServicePointTest {
           MalformedURLException {
     UUID id = UUID.randomUUID();
     createServicePoint(id, "Circ Desk 1", "cd1",
-        "Circulation Desk -- Hallway", null, 20, true);
+        "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
     CompletableFuture<Response> deleted = new CompletableFuture<>();
     send(servicePointsUrl("/" + id.toString()), HttpMethod.DELETE, null,
             SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(deleted));
@@ -217,8 +229,8 @@ public class ServicePointTest {
   // --- END TESTS --- //
 
   public static Response createServicePoint(UUID id, String name, String code,
-          String discoveryDisplayName, String description, Integer shelvingLagTime,
-      Boolean pickupLocation)
+                                            String discoveryDisplayName, String description, Integer shelvingLagTime,
+                                            Boolean pickupLocation, HoldShelfExpiryPeriod shelfExpiryPeriod)
           throws MalformedURLException, InterruptedException, ExecutionException, TimeoutException {
 
     CompletableFuture<Response> createServicePoint = new CompletableFuture<>();
@@ -231,12 +243,400 @@ public class ServicePointTest {
     if(description != null) { request.put("description", description); }
     if(shelvingLagTime != null) { request.put("shelvingLagTime", shelvingLagTime); }
     if(pickupLocation != null) { request.put("pickupLocation", pickupLocation); }
+    if(shelfExpiryPeriod != null) {request.put("holdShelfExpiryPeriod", new JsonObject(Json.encode(shelfExpiryPeriod)));}
 
     send(servicePointsUrl(""), HttpMethod.POST, request.toString(),
             SUPPORTED_CONTENT_TYPE_JSON_DEF,
             ResponseHandler.json(createServicePoint));
     return createServicePoint.get(5, TimeUnit.SECONDS);
 
+  }
+
+  @Test
+  public void canCreateServicePointWithHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 11", "cd11",
+      "Circulation Desk 11 -- Hallway", null, 20,
+      true, createHoldShelfExpiryPeriod(3, HoldShelfExpiryPeriod.DurationUnit.MINUTES));
+    assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    JsonObject responseJson = response.getJson();
+
+    assertThat(responseJson.getString("id"), notNullValue());
+    assertThat(responseJson.getString("code"), is("cd11"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 11"));
+
+    JsonObject holdShelfExpiryPeriod = responseJson.getJsonObject("holdShelfExpiryPeriod");
+    assertThat(holdShelfExpiryPeriod.getInteger("duration"), is (3));
+    assertThat(holdShelfExpiryPeriod.getString("durationUnit"), is (HoldShelfExpiryPeriod.DurationUnit.MINUTES.toString()));
+  }
+
+  @Test
+  public void canCreateServicePointWithoutPickupLocationAndHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 11", "cd11",
+      "Circulation Desk 11 -- Hallway", null, 20,
+      null, null);
+
+    assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_CREATED));
+
+    JsonObject responseJson = response.getJson();
+
+    assertThat(responseJson.getString("id"), notNullValue());
+    assertThat(responseJson.getString("code"), is("cd11"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 11"));
+  }
+
+  @Test
+  public void cannotCreateServicePointWithoutPickupLocationButWithHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 11", "cd11",
+      "Circulation Desk 11 -- Hallway", null, 20,
+      null, createHoldShelfExpiryPeriod());
+
+    assertThat(response.getStatusCode(), is(422));
+    JsonObject responseJson = response.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_BEING_PICKUP_LOC));
+  }
+
+  @Test
+  public void cannotCreateServicePointWithoutHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      true, null);
+    assertThat(response.getStatusCode(), is(422));
+
+    JsonObject responseJson = response.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_HOLD_EXPIRY));
+  }
+
+  @Test
+  public void cannotCreateServicePointWithoutBeingPickupLocation()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      false, createHoldShelfExpiryPeriod());
+    assertThat(response.getStatusCode(), is(422));
+    JsonObject responseJson = response.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_BEING_PICKUP_LOC));
+  }
+
+  @Test
+  public void cannotCreateServicePointWithoutAValidHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    Response response = createServicePoint(null, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      false, createHoldShelfExpiryPeriod(2, null));
+    assertThat(response.getStatusCode(), is(422));
+  }
+
+  @Test
+  public void canUpdateAServicePointWithHoldShelfExpiryPeriodWhenThereWasNoHoldShelfExpiryAndNotBeingPickupLocation()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      false, null);
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", true)
+      .put("holdShelfExpiryPeriod", new JsonObject(
+                                      Json.encode(createHoldShelfExpiryPeriod(5, HoldShelfExpiryPeriod.DurationUnit.WEEKS)))
+      );
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+    assertThat(updateResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    Response getResponse = getById(id);
+    JsonObject responseJson = getResponse.getJson();
+    assertThat(responseJson.getString("id"), is(id.toString()));
+    assertThat(responseJson.getString("code"), is("cd2"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 2")); //should fail
+    assertThat(responseJson.getBoolean("pickupLocation"), is(true));
+
+    JsonObject holdShelfExpiryPeriod = responseJson.getJsonObject("holdShelfExpiryPeriod");
+    assertThat(holdShelfExpiryPeriod.getInteger("duration"), is (5));
+    assertThat(holdShelfExpiryPeriod.getString("durationUnit"), is (HoldShelfExpiryPeriod.DurationUnit.WEEKS.toString()));
+  }
+
+  @Test
+  public void canUpdateAServicePointWithoutHoldShelfExpiryPeriodAndWithoutBeingPickupLocationWhenNeitherWasTheCase()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      false, null);
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", false);
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+    assertThat(updateResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    Response getResponse = getById(id);
+    JsonObject responseJson = getResponse.getJson();
+    assertThat(responseJson.getString("id"), is(id.toString()));
+    assertThat(responseJson.getString("code"), is("cd2"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 2"));
+    assertThat(responseJson.getBoolean("pickupLocation"), is(false));
+  }
+
+  @Test
+  public void cannotUpdateAServicePointWithoutHoldShelfExpiryPeriodWhenItOriginallyExisted()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20, true, createHoldShelfExpiryPeriod());
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", true);
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+
+    assertThat(updateResponse.getStatusCode(), is(422));
+    JsonObject responseJson = updateResponse.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_HOLD_EXPIRY));
+  }
+
+  @Test
+  public void cannotUpdateAServicePointWithoutHoldShelfExpiryPeriodWhenItNeverExisted()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20, false, null);
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", true);
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+
+    assertThat(updateResponse.getStatusCode(), is(422));
+    JsonObject responseJson = updateResponse.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_HOLD_EXPIRY));
+  }
+
+  @Test
+  public void cannotUpdateAServicePointWithoutBeingPickupLocationAndHoldShelfExpiryPeriodAlreadyExisted()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+
+    HoldShelfExpiryPeriod defaultHoldShelfExpiryPeriod = createHoldShelfExpiryPeriod();
+
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20, true, defaultHoldShelfExpiryPeriod);
+
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", false)
+      .put("holdShelfExpiryPeriod", new JsonObject(Json.encode(defaultHoldShelfExpiryPeriod)));
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+
+    assertThat(updateResponse.getStatusCode(), is(422));
+    JsonObject responseJson = updateResponse.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_BEING_PICKUP_LOC));
+  }
+
+  @Test
+  public void cannotUpdateAServicePointWithoutBeingPickupLocationWhileAddingHoldShelfExpiryPeriod()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+    UUID id = UUID.randomUUID();
+
+    HoldShelfExpiryPeriod defaultHoldShelfExpiryPeriod = createHoldShelfExpiryPeriod();
+
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20, false, null);
+
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", false)
+      .put("holdShelfExpiryPeriod", new JsonObject(Json.encode(defaultHoldShelfExpiryPeriod)));
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+
+    assertThat(updateResponse.getStatusCode(), is(422));
+    JsonObject responseJson = updateResponse.getJson();
+    JsonArray errorsArray = responseJson.getJsonArray("errors");
+    assertThat(errorsArray.size(), is(1));
+    JsonObject errorObject = errorsArray.getJsonObject(0);
+    assertThat(errorObject.getString("message"), is(SERVICE_POINT_CREATE_ERR_MSG_WITHOUT_BEING_PICKUP_LOC));
+  }
+
+  @Test
+  public void canUpdateAServicePointWithHoldShelfExpiryPeriodAndBeingPickupLocationWhenBothWasPresent()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      true, createHoldShelfExpiryPeriod());
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", true)
+      .put("holdShelfExpiryPeriod", new JsonObject(
+        Json.encode(createHoldShelfExpiryPeriod(5, HoldShelfExpiryPeriod.DurationUnit.WEEKS)))
+      );
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+    assertThat(updateResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    Response getResponse = getById(id);
+    JsonObject responseJson = getResponse.getJson();
+    assertThat(responseJson.getString("id"), is(id.toString()));
+    assertThat(responseJson.getString("code"), is("cd2"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 2"));
+    assertThat(responseJson.getBoolean("pickupLocation"), is(true));
+
+    JsonObject holdShelfExpiryPeriod = responseJson.getJsonObject("holdShelfExpiryPeriod");
+    assertThat(holdShelfExpiryPeriod.getInteger("duration"), is (5));
+    assertThat(holdShelfExpiryPeriod.getString("durationUnit"), is (HoldShelfExpiryPeriod.DurationUnit.WEEKS.toString()));
+  }
+
+  @Test
+  public void canUpdateAServicePointWithoutHoldShelfExpiryPeriodAndBeingPickupLocationWhenBothWasPresent()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    createServicePoint(id, "Circ Desk 1", "cd1",
+      "Circulation Desk -- Hallway", null, 20,
+      true, createHoldShelfExpiryPeriod());
+    JsonObject request = new JsonObject()
+      .put("id", id.toString())
+      .put("name", "Circ Desk 2")
+      .put("code", "cd2")
+      .put("discoveryDisplayName", "Circulation Desk -- Basement")
+      .put("pickupLocation", false);
+
+    CompletableFuture<Response> updated = new CompletableFuture<>();
+    send(servicePointsUrl("/" + id.toString()), HttpMethod.PUT, request.encode(),
+      SUPPORTED_CONTENT_TYPE_JSON_DEF, ResponseHandler.any(updated));
+
+    Response updateResponse = updated.get(5, TimeUnit.SECONDS);
+    assertThat(updateResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    Response getResponse = getById(id);
+    JsonObject responseJson = getResponse.getJson();
+    assertThat(responseJson.getString("id"), is(id.toString()));
+    assertThat(responseJson.getString("code"), is("cd2"));
+    assertThat(responseJson.getString("name"), is("Circ Desk 2"));
+    assertThat(responseJson.getBoolean("pickupLocation"), is(false));
+  }
+
+  public static HoldShelfExpiryPeriod createHoldShelfExpiryPeriod(int duration, HoldShelfExpiryPeriod.DurationUnit unit){
+      HoldShelfExpiryPeriod holdShelfExpiryPeriod = new HoldShelfExpiryPeriod();
+      holdShelfExpiryPeriod.setDuration(duration);
+      holdShelfExpiryPeriod.setDurationUnit(unit);
+      return holdShelfExpiryPeriod;
+  }
+
+  public static HoldShelfExpiryPeriod createHoldShelfExpiryPeriod(){
+      return createHoldShelfExpiryPeriod(2, HoldShelfExpiryPeriod.DurationUnit.DAYS);
   }
 
   private Response getById(UUID id)